+++ packages/analytics-js-plugins/__tests__/errorReporting/index.test.ts
@@ -1,6 +1,8 @@
 import { signal } from '@preact/signals-core';
 import { clone } from 'ramda';
 import type { IHttpClient } from '@rudderstack/analytics-js-common/types/HttpClient';
+import { defaultLogger } from '@rudderstack/analytics-js-common/__mocks__/Logger';
+import type { ExtensionPoint } from '@rudderstack/analytics-js-common/types/PluginEngine';
 import { ErrorReporting } from '../../src/errorReporting';
 
 describe('Plugin - ErrorReporting', () => {
@@ -24,6 +26,14 @@ describe('Plugin - ErrorReporting', () => {
       id: 'test-source-id',
       config: {},
     }),
+    session: {
+      sessionInfo: signal({ id: 'test-session-id' }),
+    },
+    autoTrack: {
+      pageLifecycle: {
+        visitId: signal('test-visit-id'),
+      },
+    },
   };
 
   let state: any;
@@ -34,9 +44,6 @@ describe('Plugin - ErrorReporting', () => {
   };
   const mockExtSrcLoader = {
     loadJSFile: jest.fn(() => Promise.resolve()),
-  };
-  const mockLogger = {
-    error: jest.fn(),
   };
   const mockErrReportingProviderClient = {
     notify: jest.fn(),
@@ -49,47 +56,58 @@ describe('Plugin - ErrorReporting', () => {
   });
 
   it('should add ErrorReporting plugin in the loaded plugin list', () => {
-    ErrorReporting().initialize(state);
+    ErrorReporting()?.initialize?.(state);
     expect(state.plugins.loadedPlugins.value.includes('ErrorReporting')).toBe(true);
     expect(state.reporting.isErrorReportingPluginLoaded.value).toBe(true);
     expect(state.reporting.breadcrumbs.value[0].name).toBe('Error Reporting Plugin Loaded');
   });
 
   it('should not invoke error reporting provider plugin on init if request is coming from latest core SDK', () => {
-    ErrorReporting().errorReporting.init({}, mockPluginEngine, mockExtSrcLoader, mockLogger, true);
+    (ErrorReporting()?.errorReporting as ExtensionPoint)?.init?.(
+      {},
+      mockPluginEngine,
+      mockExtSrcLoader,
+      defaultLogger,
+      true,
+    );
     expect(mockPluginEngine.invokeSingle).not.toHaveBeenCalled();
   });
 
   it('should not invoke error reporting provider plugin on init if sourceConfig do not have required parameters', () => {
-    ErrorReporting().errorReporting.init(
+    (ErrorReporting()?.errorReporting as ExtensionPoint)?.init?.(
       state,
       mockPluginEngine,
       mockExtSrcLoader,
-      mockLogger,
+      defaultLogger,
       true,
     );
     expect(mockPluginEngine.invokeSingle).not.toHaveBeenCalled();
   });
 
   it('should not invoke error reporting provider plugin on init if request is coming from old core SDK', () => {
-    ErrorReporting().errorReporting.init(state, mockPluginEngine, mockExtSrcLoader, mockLogger);
+    (ErrorReporting()?.errorReporting as ExtensionPoint)?.init?.(
+      state,
+      mockPluginEngine,
+      mockExtSrcLoader,
+      defaultLogger,
+    );
     expect(mockPluginEngine.invokeSingle).toHaveBeenCalledTimes(1);
     expect(mockPluginEngine.invokeSingle).toHaveBeenCalledWith(
       'errorReportingProvider.init',
       state,
       mockExtSrcLoader,
-      mockLogger,
+      defaultLogger,
     );
   });
 
   it('should invoke the error reporting provider plugin on notify if httpClient is not provided', () => {
     const dummyError = new Error('dummy error');
-    ErrorReporting().errorReporting.notify(
+    (ErrorReporting()?.errorReporting as ExtensionPoint)?.notify?.(
       mockPluginEngine,
       mockErrReportingProviderClient,
       dummyError,
       state,
-      mockLogger,
+      defaultLogger,
     );
     expect(mockPluginEngine.invokeSingle).toHaveBeenCalledTimes(1);
     expect(mockPluginEngine.invokeSingle).toHaveBeenCalledWith(
@@ -97,7 +115,7 @@ describe('Plugin - ErrorReporting', () => {
       mockErrReportingProviderClient,
       dummyError,
       state,
-      mockLogger,
+      defaultLogger,
     );
   });
 
@@ -119,7 +137,7 @@ describe('Plugin - ErrorReporting', () => {
         value: `The request failed due to timeout at Analytics.page (http://localhost:3001/cdn/modern/iife/rsa.js:1610:3) at RudderAnalytics.page (http://localhost:3001/cdn/modern/iife/rsa.js:1666:84)`,
       },
     });
-    ErrorReporting().errorReporting.notify(
+    (ErrorReporting()?.errorReporting as ExtensionPoint)?.notify?.(
       {},
       undefined,
       normalizedError,
@@ -154,7 +172,7 @@ describe('Plugin - ErrorReporting', () => {
         value: `ReferenceError: testUndefinedFn is not defined at Analytics.page (http://localhost:3001/cdn/modern/iife/rsa.js:1610:3) at RudderAnalytics.page (http://localhost:3001/cdn/modern/iife/rsa.js:1666:84)`,
       },
     });
-    ErrorReporting().errorReporting.notify(
+    (ErrorReporting()?.errorReporting as ExtensionPoint)?.notify?.(
       {},
       undefined,
       normalizedError,
@@ -183,7 +201,7 @@ describe('Plugin - ErrorReporting', () => {
         value: `ReferenceError: testUndefinedFn is not defined at Abcd.page (http://localhost:3001/cdn/modern/iife/abc.js:1610:3) at xyz.page (http://localhost:3001/cdn/modern/iife/abc.js:1666:84)`,
       },
     });
-    ErrorReporting().errorReporting.notify(
+    (ErrorReporting()?.errorReporting as ExtensionPoint)?.notify?.(
       {},
       undefined,
       normalizedError,
@@ -202,18 +220,24 @@ describe('Plugin - ErrorReporting', () => {
 
   it('should add a new breadcrumb', () => {
     const breadcrumbLength = state.reporting.breadcrumbs.value.length;
-    ErrorReporting().errorReporting.breadcrumb({}, undefined, 'dummy breadcrumb', undefined, state);
+    (ErrorReporting()?.errorReporting as ExtensionPoint)?.breadcrumb?.(
+      {},
+      undefined,
+      'dummy breadcrumb',
+      undefined,
+      state,
+    );
 
     expect(state.reporting.breadcrumbs.value.length).toBe(breadcrumbLength + 1);
     expect(mockPluginEngine.invokeSingle).not.toHaveBeenCalled();
   });
 
   it('should invoke the error reporting provider plugin on new breadcrumb if state is not provided', () => {
-    ErrorReporting().errorReporting.breadcrumb(
+    (ErrorReporting()?.errorReporting as ExtensionPoint)?.breadcrumb?.(
       mockPluginEngine,
       mockErrReportingProviderClient,
       'dummy breadcrumb',
-      mockLogger,
+      defaultLogger,
     );
 
     expect(mockPluginEngine.invokeSingle).toHaveBeenCalledTimes(1);
@@ -221,7 +245,7 @@ describe('Plugin - ErrorReporting', () => {
       'errorReportingProvider.breadcrumb',
       mockErrReportingProviderClient,
       'dummy breadcrumb',
-      mockLogger,
+      defaultLogger,
     );
   });
 });