+++ packages/analytics-js-plugins/__tests__/ketchConsentManager/index.test.ts
@@ -1,7 +1,8 @@
-import { state, resetState } from '@rudderstack/analytics-js/state';
-import { defaultPluginEngine } from '@rudderstack/analytics-js/services/PluginEngine';
-import { PluginsManager } from '@rudderstack/analytics-js/components/pluginsManager';
-import { StoreManager } from '@rudderstack/analytics-js/services/StoreManager/StoreManager';
+import { defaultStoreManager } from '@rudderstack/analytics-js-common/__mocks__/StoreManager';
+import { defaultLogger } from '@rudderstack/analytics-js-common/__mocks__/Logger';
+import { defaultErrorHandler } from '@rudderstack/analytics-js-common/__mocks__/ErrorHandler';
+import type { ExtensionPoint } from '@rudderstack/analytics-js-common/types/PluginEngine';
+import { resetState, state } from '../../__mocks__/state';
 import { KetchConsentManager } from '../../src/ketchConsentManager';
 
 describe('Plugin - KetchConsentManager', () => {
@@ -18,24 +19,14 @@ describe('Plugin - KetchConsentManager', () => {
     });
   });
 
-  const mockLogger = {
-    error: jest.fn(),
-    warn: jest.fn(),
-    log: jest.fn(),
-  };
-
-  const mockErrorHandler = {
-    onError: jest.fn(),
-  };
-
   it('should add KetchConsentManager plugin in the loaded plugin list', () => {
-    KetchConsentManager().initialize(state);
+    KetchConsentManager()?.initialize?.(state);
     expect(state.plugins.loadedPlugins.value.includes('KetchConsentManager')).toBe(true);
   });
 
   it('should initialize the plugin if ketch consent data is already available on the window object', () => {
     // Initialize the plugin
-    KetchConsentManager().consentManager.init(state, mockLogger);
+    (KetchConsentManager()?.consentManager as ExtensionPoint).init?.(state, defaultLogger);
 
     expect((window as any).getKetchUserConsentedPurposes).toEqual(expect.any(Function));
     expect((window as any).getKetchUserDeniedPurposes).toEqual(expect.any(Function));
@@ -53,10 +44,14 @@ describe('Plugin - KetchConsentManager', () => {
     };
 
     // Initialize the plugin
-    KetchConsentManager().consentManager.init(state, mockLogger);
+    (KetchConsentManager()?.consentManager as ExtensionPoint).init?.(state, defaultLogger);
 
     // Update the state with the consent data
-    KetchConsentManager().consentManager.updateConsentsInfo(state, undefined, mockLogger);
+    (KetchConsentManager()?.consentManager as ExtensionPoint).updateConsentsInfo?.(
+      state,
+      undefined,
+      defaultLogger,
+    );
 
     expect(state.consents.initialized.value).toBe(true);
     expect(state.consents.data.value).toStrictEqual({
@@ -74,7 +69,7 @@ describe('Plugin - KetchConsentManager', () => {
 
   it('should return undefined values when the window callbacks are invoked and there is no data in the state', () => {
     // Initialize the plugin
-    KetchConsentManager().consentManager.init(state, mockLogger);
+    (KetchConsentManager()?.consentManager as ExtensionPoint).init?.(state, defaultLogger);
 
     expect((window as any).getKetchUserConsentedPurposes()).toStrictEqual(undefined);
     expect((window as any).getKetchUserDeniedPurposes()).toStrictEqual(undefined);
@@ -91,7 +86,7 @@ describe('Plugin - KetchConsentManager', () => {
     };
 
     // Initialize the plugin
-    KetchConsentManager().consentManager.init(state, mockLogger);
+    (KetchConsentManager()?.consentManager as ExtensionPoint).init?.(state, defaultLogger);
 
     // Call the callback function
     (window as any).updateKetchConsent({
@@ -138,14 +133,15 @@ describe('Plugin - KetchConsentManager', () => {
     // Mock the ketch cookies
     document.cookie = `_ketch_consent_v1_=${window.btoa(ketchConsentString)};`;
 
-    const pluginsManager = new PluginsManager(defaultPluginEngine, undefined, mockLogger);
-    const storeManager = new StoreManager(pluginsManager, undefined, mockLogger);
-
     // Initialize the plugin
-    KetchConsentManager().consentManager.init(state, mockLogger);
+    (KetchConsentManager()?.consentManager as ExtensionPoint).init?.(state, defaultLogger);
 
     // Update the state with the consent data
-    KetchConsentManager().consentManager.updateConsentsInfo(state, storeManager, mockLogger);
+    (KetchConsentManager()?.consentManager as ExtensionPoint).updateConsentsInfo?.(
+      state,
+      defaultStoreManager,
+      defaultLogger,
+    );
 
     expect(state.consents.initialized.value).toBe(true);
     expect(state.consents.data.value).toStrictEqual({
@@ -163,11 +159,11 @@ describe('Plugin - KetchConsentManager', () => {
 
   it('should return true if the consent manager is not initialized', () => {
     expect(
-      KetchConsentManager().consentManager.isDestinationConsented(
+      (KetchConsentManager()?.consentManager as ExtensionPoint).isDestinationConsented?.(
         state,
         undefined,
-        mockErrorHandler,
-        mockLogger,
+        defaultErrorHandler,
+        defaultLogger,
       ),
     ).toBe(true);
   });
@@ -186,11 +182,11 @@ describe('Plugin - KetchConsentManager', () => {
     };
 
     expect(
-      KetchConsentManager().consentManager.isDestinationConsented(
+      (KetchConsentManager()?.consentManager as ExtensionPoint).isDestinationConsented?.(
         state,
         destConfig,
-        mockErrorHandler,
-        mockLogger,
+        defaultErrorHandler,
+        defaultLogger,
       ),
     ).toBe(true);
   });
@@ -210,11 +206,11 @@ describe('Plugin - KetchConsentManager', () => {
     };
 
     expect(
-      KetchConsentManager().consentManager.isDestinationConsented(
+      (KetchConsentManager()?.consentManager as ExtensionPoint).isDestinationConsented?.(
         state,
         destConfig,
-        mockErrorHandler,
-        mockLogger,
+        defaultErrorHandler,
+        defaultLogger,
       ),
     ).toBe(true);
   });
@@ -241,11 +237,11 @@ describe('Plugin - KetchConsentManager', () => {
     };
 
     expect(
-      KetchConsentManager().consentManager.isDestinationConsented(
+      (KetchConsentManager()?.consentManager as ExtensionPoint).isDestinationConsented?.(
         state,
         destConfig,
-        mockErrorHandler,
-        mockLogger,
+        defaultErrorHandler,
+        defaultLogger,
       ),
     ).toBe(true);
   });
@@ -272,11 +268,11 @@ describe('Plugin - KetchConsentManager', () => {
     };
 
     expect(
-      KetchConsentManager().consentManager.isDestinationConsented(
+      (KetchConsentManager()?.consentManager as ExtensionPoint).isDestinationConsented?.(
         state,
         destConfig,
-        mockErrorHandler,
-        mockLogger,
+        defaultErrorHandler,
+        defaultLogger,
       ),
     ).toBe(false);
   });
@@ -284,6 +280,7 @@ describe('Plugin - KetchConsentManager', () => {
   it('should return true and log an error if any exception is thrown while checking if the destination is consented', () => {
     state.consents.initialized.value = true;
     state.consents.data.value = {
+      // @ts-expect-error Intentionally set to null to throw an exception
       allowedConsentIds: null, // This will throw an exception
       deniedConsentIds: ['purpose2', 'purpose4'],
     };
@@ -303,14 +300,14 @@ describe('Plugin - KetchConsentManager', () => {
     };
 
     expect(
-      KetchConsentManager().consentManager.isDestinationConsented(
+      (KetchConsentManager()?.consentManager as ExtensionPoint).isDestinationConsented?.(
         state,
         destConfig,
-        mockErrorHandler,
-        mockLogger,
+        defaultErrorHandler,
+        defaultLogger,
       ),
     ).toBe(true);
-    expect(mockErrorHandler.onError).toHaveBeenCalledWith(
+    expect(defaultErrorHandler.onError).toHaveBeenCalledWith(
       new TypeError("Cannot read properties of null (reading 'includes')"),
       'KetchConsentManagerPlugin',
       'Failed to determine the consent status for the destination. Please check the destination configuration and try again.',
@@ -350,12 +347,9 @@ describe('Plugin - KetchConsentManager', () => {
         },
       ],
     };
-    const isDestinationConsented = KetchConsentManager().consentManager.isDestinationConsented(
-      state,
-      destConfig,
-      mockErrorHandler,
-      mockLogger,
-    );
+    const isDestinationConsented = (
+      KetchConsentManager()?.consentManager as ExtensionPoint
+    ).isDestinationConsented?.(state, destConfig, defaultErrorHandler, defaultLogger);
     expect(isDestinationConsented).toBe(false);
   });
 
@@ -381,12 +375,9 @@ describe('Plugin - KetchConsentManager', () => {
         },
       ],
     };
-    const isDestinationConsented = KetchConsentManager().consentManager.isDestinationConsented(
-      state,
-      destConfig,
-      mockErrorHandler,
-      mockLogger,
-    );
+    const isDestinationConsented = (
+      KetchConsentManager()?.consentManager as ExtensionPoint
+    ).isDestinationConsented?.(state, destConfig, defaultErrorHandler, defaultLogger);
     expect(isDestinationConsented).toBe(true);
   });
 
@@ -423,12 +414,9 @@ describe('Plugin - KetchConsentManager', () => {
         },
       ],
     };
-    const isDestinationConsented = KetchConsentManager().consentManager.isDestinationConsented(
-      state,
-      destConfig,
-      mockErrorHandler,
-      mockLogger,
-    );
+    const isDestinationConsented = (
+      KetchConsentManager()?.consentManager as ExtensionPoint
+    ).isDestinationConsented?.(state, destConfig, defaultErrorHandler, defaultLogger);
     expect(isDestinationConsented).toBe(true);
   });
 
@@ -455,11 +443,11 @@ describe('Plugin - KetchConsentManager', () => {
       ],
     };
     expect(
-      KetchConsentManager().consentManager.isDestinationConsented(
+      (KetchConsentManager()?.consentManager as ExtensionPoint).isDestinationConsented?.(
         state,
         destConfig,
-        mockErrorHandler,
-        mockLogger,
+        defaultErrorHandler,
+        defaultLogger,
       ),
     ).toBe(true);
   });
@@ -467,7 +455,7 @@ describe('Plugin - KetchConsentManager', () => {
   it('should return appropriate value when the resolution strategy not set', () => {
     state.consents.initialized.value = true;
     state.consents.provider.value = 'ketch';
-    state.consents.resolutionStrategy.value = null;
+    state.consents.resolutionStrategy.value = undefined;
     state.consents.data.value = {
       allowedConsentIds: ['C0001', 'C0002', 'C0003'],
     };
@@ -487,11 +475,11 @@ describe('Plugin - KetchConsentManager', () => {
       ],
     };
     expect(
-      KetchConsentManager().consentManager.isDestinationConsented(
+      (KetchConsentManager()?.consentManager as ExtensionPoint).isDestinationConsented?.(
         state,
         destConfig,
-        mockErrorHandler,
-        mockLogger,
+        defaultErrorHandler,
+        defaultLogger,
       ),
     ).toBe(true);
   });