+++ packages/analytics-js-plugins/__tests__/xhrQueue/index.test.ts
@@ -1,18 +1,17 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import { batch } from '@preact/signals-core';
-import { HttpClient } from '@rudderstack/analytics-js/services/HttpClient';
-import { state } from '@rudderstack/analytics-js/state';
 import { mergeDeepRight } from '@rudderstack/analytics-js-common/utilities/object';
-import { PluginsManager } from '@rudderstack/analytics-js/components/pluginsManager';
-import { defaultPluginEngine } from '@rudderstack/analytics-js/services/PluginEngine';
-import { defaultErrorHandler } from '@rudderstack/analytics-js/services/ErrorHandler';
-import { defaultLogger } from '@rudderstack/analytics-js/services/Logger';
-import { StoreManager } from '@rudderstack/analytics-js/services/StoreManager';
+import { defaultStoreManager } from '@rudderstack/analytics-js-common/__mocks__/StoreManager';
 import type { RudderEvent } from '@rudderstack/analytics-js-common/types/Event';
-import type { ILogger } from '@rudderstack/analytics-js-common/types/Logger';
-import type { IHttpClient } from '@rudderstack/analytics-js-common/types/HttpClient';
+import { defaultLogger } from '@rudderstack/analytics-js-common/__mocks__/Logger';
+import { defaultErrorHandler } from '@rudderstack/analytics-js-common/__mocks__/ErrorHandler';
+import type { ExtensionPoint } from '@rudderstack/analytics-js-common/types/PluginEngine';
+import type { RetryQueue } from '../../src/utilities/retryQueue/RetryQueue';
+import type { QueueItem, QueueItemData } from '../../src/types/plugins';
+import { resetState, state } from '../../__mocks__/state';
 import { XhrQueue } from '../../src/xhrQueue';
 import { Schedule } from '../../src/utilities/retryQueue/Schedule';
+import { defaultHttpClient } from '../../__mocks__/HttpClient';
 
 jest.mock('@rudderstack/analytics-js-common/utilities/timestamp', () => ({
   ...jest.requireActual('@rudderstack/analytics-js-common/utilities/timestamp'),
@@ -25,19 +24,8 @@ jest.mock('@rudderstack/analytics-js-common/utilities/uuId', () => ({
 }));
 
 describe('XhrQueue', () => {
-  const defaultPluginsManager = new PluginsManager(
-    defaultPluginEngine,
-    defaultErrorHandler,
-    defaultLogger,
-  );
-
-  const defaultStoreManager = new StoreManager(defaultPluginsManager);
-
-  const mockLogger = {
-    error: jest.fn(),
-  } as unknown as ILogger;
-
   beforeAll(() => {
+    resetState();
     batch(() => {
       state.lifecycle.writeKey.value = 'sampleWriteKey';
       state.lifecycle.activeDataplaneUrl.value = 'https://sampleurl.com';
@@ -51,33 +39,35 @@ describe('XhrQueue', () => {
     });
   });
 
-  const httpClient = new HttpClient();
-
   it('should add itself to the loaded plugins list on initialized', () => {
-    XhrQueue().initialize(state);
+    XhrQueue()?.initialize?.(state);
 
     expect(state.plugins.loadedPlugins.value).toContain('XhrQueue');
   });
 
   it('should return a queue object on init', () => {
-    const queue = XhrQueue().dataplaneEventsQueue?.init(
+    const queue = (XhrQueue()?.dataplaneEventsQueue as ExtensionPoint).init?.(
       state,
-      httpClient,
+      defaultHttpClient,
       defaultStoreManager,
       defaultErrorHandler,
       defaultLogger,
-    );
+    ) as RetryQueue;
 
     expect(queue).toBeDefined();
     expect(queue.name).toBe('rudder_sampleWriteKey');
   });
 
   it('should add item in queue on enqueue', () => {
-    const queue = XhrQueue().dataplaneEventsQueue?.init(state, httpClient, defaultStoreManager);
+    const queue = (XhrQueue()?.dataplaneEventsQueue as ExtensionPoint).init?.(
+      state,
+      defaultHttpClient,
+      defaultStoreManager,
+    ) as RetryQueue;
 
     const addItemSpy = jest.spyOn(queue, 'addItem');
 
-    const event: RudderEvent = {
+    const event = {
       type: 'track',
       event: 'test',
       userId: 'test',
@@ -87,11 +77,11 @@ describe('XhrQueue', () => {
       anonymousId: 'sampleAnonId',
       messageId: 'test',
       originalTimestamp: 'test',
-    };
+    } as unknown as RudderEvent;
 
-    XhrQueue().dataplaneEventsQueue?.enqueue(state, queue, event);
+    (XhrQueue()?.dataplaneEventsQueue as ExtensionPoint).enqueue?.(state, queue, event);
 
-    expect(addItemSpy).toBeCalledWith({
+    expect(addItemSpy).toHaveBeenCalledWith({
       url: 'https://sampleurl.com/v1/track',
       headers: {
         AnonymousId: 'c2FtcGxlQW5vbklk', // Base64 encoded anonymousId
@@ -103,15 +93,18 @@ describe('XhrQueue', () => {
   });
 
   it('should process queue item on start', () => {
-    const mockHttpClient = {
-      getAsyncData: ({ callback }) => {
-        callback(true);
-      },
-      setAuthHeader: jest.fn(),
-    };
-    const queue = XhrQueue().dataplaneEventsQueue?.init(state, mockHttpClient, defaultStoreManager);
+    // Mock getAsyncData to return a successful response
 
-    const event: RudderEvent = {
+    defaultHttpClient.getAsyncData.mockImplementation(({ callback }) => {
+      callback?.(true);
+    });
+    const queue = (XhrQueue()?.dataplaneEventsQueue as ExtensionPoint).init?.(
+      state,
+      defaultHttpClient,
+      defaultStoreManager,
+    ) as RetryQueue;
+
+    const event = {
       type: 'track',
       event: 'test',
       userId: 'test',
@@ -121,17 +114,17 @@ describe('XhrQueue', () => {
       anonymousId: 'sampleAnonId',
       messageId: 'test',
       originalTimestamp: 'test',
-    };
+    } as unknown as RudderEvent;
 
     const queueProcessCbSpy = jest.spyOn(queue, 'processQueueCb');
 
-    XhrQueue().dataplaneEventsQueue?.enqueue(state, queue, event);
+    (XhrQueue()?.dataplaneEventsQueue as ExtensionPoint).enqueue?.(state, queue, event);
 
     // Explicitly start the queue to process the item
     // In actual implementation, this is done based on the state signals
     queue.start();
 
-    expect(queueProcessCbSpy).toBeCalledWith(
+    expect(queueProcessCbSpy).toHaveBeenCalledWith(
       {
         url: 'https://sampleurl.com/v1/track',
         headers: {
@@ -146,26 +139,24 @@ describe('XhrQueue', () => {
     );
 
     // Item is successfully processed and removed from queue
-    expect(queue.getStorageEntry('queue').length).toBe(0);
+    expect((queue.getStorageEntry('queue') as QueueItem<QueueItemData>[]).length).toBe(0);
 
     queueProcessCbSpy.mockRestore();
   });
 
   it('should log error on retryable failure and requeue the item', () => {
-    const mockHttpClient = {
-      getAsyncData: ({ callback }) => {
-        callback(false, { error: 'some error', xhr: { status: 429 } });
-      },
-      setAuthHeader: jest.fn(),
-    } as unknown as IHttpClient;
+    // Mock getAsyncData to return a retryable failure
 
-    const queue = XhrQueue().dataplaneEventsQueue?.init(
+    defaultHttpClient.getAsyncData.mockImplementation(({ callback }) => {
+      callback?.(false, { error: 'some error', xhr: { status: 429 } });
+    });
+    const queue = (XhrQueue()?.dataplaneEventsQueue as ExtensionPoint).init?.(
       state,
-      mockHttpClient,
+      defaultHttpClient,
       defaultStoreManager,
       undefined,
-      mockLogger,
-    );
+      defaultLogger,
+    ) as RetryQueue;
 
     const schedule = new Schedule();
     // Override the timestamp generation function to return a fixed value
@@ -173,7 +164,7 @@ describe('XhrQueue', () => {
 
     queue.schedule = schedule;
 
-    const event: RudderEvent = {
+    const event = {
       type: 'track',
       event: 'test',
       userId: 'test',
@@ -183,15 +174,15 @@ describe('XhrQueue', () => {
       anonymousId: 'sampleAnonId',
       messageId: 'test',
       originalTimestamp: 'test',
-    };
+    } as unknown as RudderEvent;
 
-    XhrQueue().dataplaneEventsQueue?.enqueue(state, queue, event);
+    (XhrQueue()?.dataplaneEventsQueue as ExtensionPoint).enqueue?.(state, queue, event);
 
     // Explicitly start the queue to process the item
     // In actual implementation, this is done based on the state signals
     queue.start();
 
-    expect(mockLogger.error).toBeCalledWith(
+    expect(defaultLogger.error).toHaveBeenCalledWith(
       'XhrQueuePlugin:: Failed to deliver event(s) to https://sampleurl.com/v1/track. It/they will be retried.',
     );
 
@@ -229,18 +220,13 @@ describe('XhrQueue', () => {
       };
     });
 
-    const mockHttpClient = {
-      getAsyncData: jest.fn(),
-      setAuthHeader: jest.fn(),
-    } as unknown as IHttpClient;
-
-    const queue = XhrQueue().dataplaneEventsQueue?.init(
+    const queue = (XhrQueue()?.dataplaneEventsQueue as ExtensionPoint).init?.(
       state,
-      mockHttpClient,
+      defaultHttpClient,
       defaultStoreManager,
       undefined,
-      mockLogger,
-    );
+      defaultLogger,
+    ) as RetryQueue;
     const queueProcessCbSpy = jest.spyOn(queue, 'processQueueCb');
 
     const schedule = new Schedule();
@@ -249,7 +235,7 @@ describe('XhrQueue', () => {
 
     queue.schedule = schedule;
 
-    const event: RudderEvent = {
+    const event = {
       type: 'track',
       event: 'test',
       userId: 'test',
@@ -259,9 +245,9 @@ describe('XhrQueue', () => {
       anonymousId: 'sampleAnonId',
       messageId: 'test',
       originalTimestamp: 'test',
-    };
+    } as unknown as RudderEvent;
 
-    const event2: RudderEvent = {
+    const event2 = {
       type: 'track',
       event: 'test2',
       userId: 'test2',
@@ -271,16 +257,16 @@ describe('XhrQueue', () => {
       anonymousId: 'sampleAnonId',
       messageId: 'test2',
       originalTimestamp: 'test2',
-    };
+    } as unknown as RudderEvent;
 
-    XhrQueue().dataplaneEventsQueue?.enqueue(state, queue, event);
-    XhrQueue().dataplaneEventsQueue?.enqueue(state, queue, event2);
+    (XhrQueue()?.dataplaneEventsQueue as ExtensionPoint).enqueue?.(state, queue, event);
+    (XhrQueue()?.dataplaneEventsQueue as ExtensionPoint).enqueue?.(state, queue, event2);
 
     // Explicitly start the queue to process the item
     // In actual implementation, this is done based on the state signals
     queue.start();
 
-    expect(queueProcessCbSpy).toBeCalledWith(
+    expect(queueProcessCbSpy).toHaveBeenCalledWith(
       [
         {
           url: 'https://sampleurl.com/v1/track',
@@ -303,7 +289,7 @@ describe('XhrQueue', () => {
       true,
     );
 
-    expect(mockHttpClient.getAsyncData).toBeCalledWith({
+    expect(defaultHttpClient.getAsyncData).toHaveBeenCalledWith({
       url: 'https://sampleurl.com/v1/batch',
       options: {
         method: 'POST',