+++ packages/analytics-js-service-worker/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.2.17](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.2.16...@rudderstack/analytics-js-service-worker@3.2.17) (2024-12-17)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.14.14`
+## [3.2.16](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.2.15...@rudderstack/analytics-js-service-worker@3.2.16) (2024-12-06)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.14.13`
 ## [3.2.15](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.2.14...@rudderstack/analytics-js-service-worker@3.2.15) (2024-11-22)
 
 ### Dependency Updates